refactor bone edit helpers
