+++ io_scene_xray/edit_helpers/bone_center.py
@@ -3,10 +3,10 @@ import mathutils
 
 from io_scene_xray import registry
 from io_scene_xray.xray_motions import MATRIX_BONE, MATRIX_BONE_INVERTED
-from .base import AbstractHelper
+from .base_bone import AbstractBoneEditHelper
 
 
-class _BoneCenterEditHelper(AbstractHelper):
+class _BoneCenterEditHelper(AbstractBoneEditHelper):
     def draw(self, layout, context):
         if self.is_active(context):
             layout.operator(_CenterEditApplyOp.bl_idname, icon='FILE_TICK')
@@ -19,26 +19,6 @@ class _BoneCenterEditHelper(AbstractHelper):
             lay.enabled = False
         lay.operator(EditCenter.bl_idname, text='Edit Center')
 
-    def _is_active_target(self, target, context):
-        if target is None:
-            return False
-        bone = context.active_bone
-        if bone is None:
-            return False
-        return (bone.name == target.name) and (bone.id_data == target.id_data)
-
-    def _get_target_object(self, helper):
-        split = helper.xray.helper_data.split('/')
-        if len(split) != 2:
-            return
-        arm = bpy.data.armatures.get(split[0], None)
-        if arm is None:
-            return
-        bone = arm.bones.get(split[1], None)
-        if bone is None:
-            return
-        return bone
-
     def _create_helper(self, name):
         helper = bpy.data.objects.new(name, None)
         helper.empty_draw_size = 0.05
@@ -46,12 +26,12 @@ class _BoneCenterEditHelper(AbstractHelper):
         return helper
 
     def _update_helper(self, helper, target):
+        super()._update_helper(helper, target)
         bone = target
 
         mat = bone.matrix_local * MATRIX_BONE_INVERTED
         mat *= mathutils.Matrix.Translation(bone.xray.mass.center)
         helper.location = mat.to_translation()
-        helper.xray.helper_data = bone.id_data.name + '/' + bone.name
 
 
 HELPER = _BoneCenterEditHelper('bone-center-edit')
